requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
amqp==5.3.1
    # via kombu
annotated-types==0.7.0
    # via pydantic
anthropic==0.78.0
    # via -r requirements.in
anyio==4.12.1
    # via
    #   anthropic
    #   httpx
asgiref==3.11.1
    # via
    #   django
    #   django-htmx
billiard==4.2.4
    # via celery
black==26.1.0
    # via -r requirements.in
boto3==1.42.43
    # via django-storages
botocore==1.42.43
    # via
    #   boto3
    #   s3transfer
build==1.4.0
    # via pip-tools
celery[redis]==5.6.2
    # via
    #   -r requirements.in
    #   django-celery-beat
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
click==8.3.1
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
coverage==7.13.3
    # via -r requirements.in
cron-descriptor==2.0.6
    # via django-celery-beat
distro==1.9.0
    # via anthropic
django==5.2.11
    # via
    #   -r requirements.in
    #   django-celery-beat
    #   django-htmx
    #   django-storages
    #   django-timezone-field
    #   django-unfold
django-celery-beat==2.8.1
    # via -r requirements.in
django-htmx==1.27.0
    # via -r requirements.in
django-ratelimit==4.1.0
    # via -r requirements.in
django-storages[boto3]==1.14.6
    # via -r requirements.in
django-timezone-field==7.2.1
    # via django-celery-beat
django-unfold==0.78.1
    # via -r requirements.in
docstring-parser==0.17.0
    # via anthropic
et-xmlfile==2.0.0
    # via openpyxl
factory-boy==3.3.3
    # via -r requirements.in
faker==40.1.2
    # via factory-boy
flake8==7.3.0
    # via -r requirements.in
gunicorn==25.0.2
    # via -r requirements.in
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via anthropic
idna==3.11
    # via
    #   anyio
    #   httpx
iniconfig==2.3.0
    # via pytest
isort==7.0.0
    # via -r requirements.in
jiter==0.13.0
    # via anthropic
jmespath==1.1.0
    # via
    #   boto3
    #   botocore
kombu[redis]==5.6.2
    # via celery
mccabe==0.7.0
    # via flake8
mypy-extensions==1.1.0
    # via black
openpyxl==3.1.5
    # via -r requirements.in
packaging==26.0
    # via
    #   black
    #   build
    #   gunicorn
    #   kombu
    #   pytest
    #   wheel
pathspec==1.0.4
    # via black
pi-heif==1.2.0
    # via -r requirements.in
pillow==12.1.0
    # via
    #   -r requirements.in
    #   pi-heif
    #   python-barcode
    #   qrcode
pip-tools==7.5.2
    # via -r requirements.in
platformdirs==4.5.1
    # via black
pluggy==1.6.0
    # via pytest
prompt-toolkit==3.0.52
    # via click-repl
psycopg[binary]==3.3.2
    # via -r requirements.in
psycopg-binary==3.3.2
    # via psycopg
pycodestyle==2.14.0
    # via flake8
pydantic==2.12.5
    # via anthropic
pydantic-core==2.41.5
    # via pydantic
pyflakes==3.4.0
    # via flake8
pygments==2.19.2
    # via pytest
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==9.0.2
    # via pytest-django
pytest-django==4.11.1
    # via -r requirements.in
python-barcode[images]==0.16.1
    # via -r requirements.in
python-crontab==3.3.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   celery
pytokens==0.4.1
    # via black
qrcode[pil]==8.2
    # via -r requirements.in
redis==6.4.0
    # via
    #   -r requirements.in
    #   kombu
s3transfer==0.16.0
    # via boto3
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anthropic
sqlparse==0.5.5
    # via django
typing-extensions==4.15.0
    # via
    #   anthropic
    #   cron-descriptor
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via
    #   django-celery-beat
    #   kombu
tzlocal==5.3.1
    # via celery
urllib3==2.6.3
    # via botocore
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.5.3
    # via prompt-toolkit
wheel==0.46.3
    # via pip-tools
whitenoise==6.11.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
